#4538: remove dead code
